# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Platform Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
  rememberMe?: boolean;
}

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  disabled: 'disabled',
} as const;

export interface User {
  id: string;
  username: string;
  fullName: string;
  email: string;
  role: string;
  status: UserStatus;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserInput {
  username: string;
  fullName: string;
  email: string;
  role: string;
  password?: string;
}

export interface UserUpdate {
  fullName?: string;
  email?: string;
  role?: string;
  password?: string;
}

export interface ToggleUserInput {
  enabled: boolean;
}

export interface Role {
  name: string;
  description: string;
  permissions: string[];
}

export type ServiceStatusStatus = typeof ServiceStatusStatus[keyof typeof ServiceStatusStatus];


export const ServiceStatusStatus = {
  Running: 'Running',
  Stopped: 'Stopped',
  Unknown: 'Unknown',
} as const;

export interface ServiceStatus {
  name: string;
  status: ServiceStatusStatus;
  /** @nullable */
  description?: string | null;
}

export interface AuditLog {
  id: string;
  actor: string;
  action: string;
  timestamp: string;
  /** @nullable */
  details?: string | null;
}

export interface DashboardSummary {
  services: ServiceStatus[];
  clusterStatus: string;
  workerCount: number;
  recentActivity: AuditLog[];
}

export interface SparkCluster {
  clusterName: string;
  namespace: string;
  masterStatus: string;
  workerStatus: string;
  currentWorkerReplicas: number;
  desiredWorkerReplicas: number;
  sparkMasterUrl: string;
  sparkUiUrl: string;
}

export interface ScaleInput {
  workerCount: number;
}

export interface ActionResult {
  success: boolean;
  message: string;
}

export type SparkPodRole = typeof SparkPodRole[keyof typeof SparkPodRole];


export const SparkPodRole = {
  master: 'master',
  worker: 'worker',
} as const;

export interface SparkPod {
  podName: string;
  role: SparkPodRole;
  status: string;
  node: string;
  cpu: string;
  memory: string;
  age: string;
}

export interface ClusterEvent {
  type: string;
  reason: string;
  message: string;
  timestamp: string;
}

export type PlatformServiceStatus = typeof PlatformServiceStatus[keyof typeof PlatformServiceStatus];


export const PlatformServiceStatus = {
  Running: 'Running',
  Stopped: 'Stopped',
  Unknown: 'Unknown',
} as const;

export interface PlatformService {
  id: string;
  name: string;
  description: string;
  namespace: string;
  status: PlatformServiceStatus;
  url: string;
  category: string;
  /** @nullable */
  isJdbc?: boolean | null;
}

export type SparkClusterProfileSize = typeof SparkClusterProfileSize[keyof typeof SparkClusterProfileSize];


export const SparkClusterProfileSize = {
  small: 'small',
  medium: 'medium',
  large: 'large',
} as const;

export type SparkClusterProfileStatus = typeof SparkClusterProfileStatus[keyof typeof SparkClusterProfileStatus];


export const SparkClusterProfileStatus = {
  RUNNING: 'RUNNING',
  STOPPED: 'STOPPED',
  SCALING: 'SCALING',
  UNKNOWN: 'UNKNOWN',
} as const;

export interface SparkClusterProfile {
  clusterName: string;
  namespace: string;
  size: SparkClusterProfileSize;
  status: SparkClusterProfileStatus;
  currentWorkerReplicas: number;
  desiredWorkerReplicas: number;
  defaultWorkerReplicas: number;
  minWorkers: number;
  maxWorkers: number;
  driverCpu: string;
  driverMemory: string;
  workerCpu: string;
  workerMemory: string;
  sparkMasterUrl: string;
  sparkUiUrl: string;
  /** @nullable */
  executorMemory?: string | null;
  /** @nullable */
  executorCores?: number | null;
  /** @nullable */
  shufflePartitions?: number | null;
  /** @nullable */
  dynamicAllocationEnabled?: boolean | null;
}

export interface SparkClusterListResponse {
  clusters: SparkClusterProfile[];
}

export interface SparkClusterFull {
  cluster: SparkClusterProfile;
  pods?: SparkPod[];
}

export interface ScaleClusterInput {
  workers: number;
}

export interface ClusterConfigInput {
  driverCpu?: string;
  driverMemory?: string;
  workerCpu?: string;
  workerMemory?: string;
  executorMemory?: string;
  executorCores?: number;
  shufflePartitions?: number;
  dynamicAllocationEnabled?: boolean;
  desiredWorkerReplicas?: number;
}