# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Data Platform Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User deleted
  /users/{id}/toggle:
    post:
      operationId: toggleUserStatus
      tags: [users]
      summary: Enable or disable user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleUserInput"
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, fullName, email, role, status, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [active, disabled]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, fullName, email, role]
      properties:
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        password:
          type: string
    ToggleUserInput:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
...