# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Platform Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  fullName?: string;
  email?: string;
  role?: string;
  password?: string;
}